Bump debhelper compat level to 9
10 is the current version, but jessie only has debhelper 9,
so leave our options open for potential backport.

Apparently, if we used dh_binary, we would not need to have
dependencies on the binary-* targets, but since we do not use
dh_binary, leave things in place.

Rewriting the rules file to make better use of modern debhelper
features is left for future work.