Move to Priority: optional
Now that policy 4.0.1 allows priority-standard packages to depend on
priority-optional packages, we no longer need to make our library
packages priority-standard so that they can be depended on by other
packages that are part of the standard install.

In fact, since all of our non-library packages already overrode the
default to be "optional", all of our binary packages are priority
optional, so change the default and remove the overrides from the
binary package stanzas.