Initial work at moving to dh
First, get dh_auto_configure and dh_auto_build working.

This may produce a package but such a package is known not to be valid
among other reasons because we don't call dh_installsystemd even
though we're claiming to be dh 12 compat.

Once we actually move to using dh this will happen.