Drop obsolete dependency on package lsb-base
The package is now empty and no longer used. This fixes Lintian errors:

    E: krb5-admin-server: depends-on-obsolete-package Depends: lsb-base (>= 3.0-6)
    E: krb5-kdc: depends-on-obsolete-package Depends: lsb-base (>= 3.0-6)
    E: krb5-kpropd: depends-on-obsolete-package Depends: lsb-base

The functionality of lsb-base is in the Essential:yes set since
Bullseye. The package itself is now an empty transitional package
(because debootstrap doesn't understand the Provides relationship) which
depends on the new provider of the functionality, sysvinit-utils, which
is also in the Essential:yes set.

More information in
https://lists.debian.org/debian-devel/2023/01/msg00150.html
and https://tracker.debian.org/media/packages/l/lsb/changelog-11.5